requirements.txt
connexion == 2.6.0
connexion[swagger-ui]
python_dateutil == 2.6.0
setuptools >= 21.0.0

Flask~=1.1.2
six~=1.15.0
Werkzeug~=1.0.1
pytz~=2021.1
mongoengine~=0.23.0
flask-mongoengine==1.0.0